+++ lib/widgets/common/action_mixins/entry_storage.dart
@@ -5,11 +5,14 @@ import 'package:aves/app_mode.dart';
 import 'package:aves/model/entry/entry.dart';
 import 'package:aves/model/entry/extensions/multipage.dart';
 import 'package:aves/model/entry/extensions/props.dart';
+import 'package:aves/model/fgw/share_copied_entry.dart';
 import 'package:aves/model/filters/album.dart';
+import 'package:aves/model/filters/path.dart';
 import 'package:aves/model/filters/trash.dart';
 import 'package:aves/model/highlight.dart';
 import 'package:aves/model/metadata/date_modifier.dart';
 import 'package:aves/model/multipage.dart';
+import 'package:aves/model/settings/enums/presentation.dart';
 import 'package:aves/model/settings/settings.dart';
 import 'package:aves/model/source/collection_lens.dart';
 import 'package:aves/model/source/collection_source.dart';
@@ -19,6 +22,7 @@ import 'package:aves/services/common/services.dart';
 import 'package:aves/services/media/enums.dart';
 import 'package:aves/services/media/media_edit_service.dart';
 import 'package:aves/theme/durations.dart';
+import 'package:aves/theme/format.dart';
 import 'package:aves/theme/themes.dart';
 import 'package:aves/utils/android_file_utils.dart';
 import 'package:aves/widgets/collection/collection_page.dart';
@@ -125,7 +129,9 @@ mixin EntryStorageMixin on FeedbackMixin, PermissionAwareMixin, SizeAwareMixin {
                         settings: const RouteSettings(name: CollectionPage.routeName),
                         builder: (context) => CollectionPage(
                           source: source,
-                          filters: {AlbumFilter(destinationAlbum, source.getAlbumDisplayName(context, destinationAlbum))},
+                          filters: {
+                            AlbumFilter(destinationAlbum, source.getAlbumDisplayName(context, destinationAlbum))
+                          },
                           highlightTest: (entry) => newUris.contains(entry.uri),
                         ),
                       ),
@@ -157,19 +163,54 @@ mixin EntryStorageMixin on FeedbackMixin, PermissionAwareMixin, SizeAwareMixin {
     transientMultiPageInfo.forEach((v) => v.dispose());
   }
 
+  Future<void> deleteExpiredShareCopied(BuildContext context) async {
+    final source = context.read<CollectionSource>();
+    //await shareCopiedEntries.init();
+    final todoEntries = source.allEntries
+        .where((entry) => {PathFilter(androidFileUtils.avesShareByCopyPath)}.any((f) => f.test(entry)))
+        .toSet();
+    final expiredEntries = todoEntries.where(shareCopiedEntries.isExpiredCopied).toSet();
+    if (expiredEntries.isEmpty) return;
+    debugPrint('$runtimeType deleteExpiredShareCopied\n'
+        'todoEntries $todoEntries \n'
+        'expiredEntries $expiredEntries\n');
+
+    final entriesByDestination = <String, Set<AvesEntry>>{};
+    entriesByDestination[AndroidFileUtils.trashDirPath] = expiredEntries;
+
+    debugPrint('$runtimeType deleteExpiredShareCopied\n'
+        'settings.enableBin ${settings.enableBin} \n'
+        ' settings.shareByCopyExpiredRemoveUseBin ${settings.shareByCopyExpiredRemoveUseBin}\n');
+    source.pauseMonitoring();
+    final useBin = settings.enableBin && settings.shareByCopyExpiredRemoveUseBin;
+    debugPrint('$runtimeType deleteExpiredShareCopied delete to useBin $useBin \n');
+    // delete forever trashed items
+    await EntrySetActionDelegate().doDelete(
+      context: context,
+      entries: expiredEntries,
+      enableBin: useBin,
+      isShareByCopyDelete: true,
+    );
+    await shareCopiedEntries.removeEntries(expiredEntries);
+    //await delegate.setDateToNow(context,entries:todoEntries,showConfirm: false);
+    return;
+  }
+
   Future<void> doQuickMove(
     BuildContext context, {
     required MoveType moveType,
     required Map<String, Iterable<AvesEntry>> entriesByDestination,
     bool hideShowAction = false,
     VoidCallback? onSuccess,
+    bool isShareByCopyDelete = false,
   }) async {
+    Set<AvesEntry> updatedEntries = {};
     final entries = entriesByDestination.values.expand((v) => v).toSet();
     final todoCount = entries.length;
     assert(todoCount > 0);
 
     final toBin = moveType == MoveType.toBin;
-    final copy = moveType == MoveType.copy;
+    final copy = moveType == MoveType.copy || moveType == MoveType.shareByCopy;
 
     // permission for modification at destinations
     final destinationAlbums = entriesByDestination.keys.toSet();
@@ -177,7 +218,8 @@ mixin EntryStorageMixin on FeedbackMixin, PermissionAwareMixin, SizeAwareMixin {
 
     // permission for modification at origins
     final originAlbums = entries.map((e) => e.directory).whereNotNull().toSet();
-    if ({MoveType.move, MoveType.toBin}.contains(moveType) && !await checkStoragePermissionForAlbums(context, originAlbums, entries: entries)) return;
+    if ({MoveType.move, MoveType.toBin}.contains(moveType) &&
+        !await checkStoragePermissionForAlbums(context, originAlbums, entries: entries)) return;
 
     final hasEnoughSpaceByDestination = await Future.wait(destinationAlbums.map((destinationAlbum) {
       return checkFreeSpaceForMove(context, entries, destinationAlbum, moveType);
@@ -192,17 +234,21 @@ mixin EntryStorageMixin on FeedbackMixin, PermissionAwareMixin, SizeAwareMixin {
         ...entries.map((v) => '${v.filenameWithoutExtension}${v.extension}'),
         // do not guard up front based on directory existence,
         // as conflicts could be within moved entries scattered across multiple albums
-        if (await destinationDirectory.exists()) ...destinationDirectory.listSync().map((v) => pContext.basename(v.path)),
+        if (await destinationDirectory.exists())
+          ...destinationDirectory.listSync().map((v) => pContext.basename(v.path)),
       ].map((v) => v.toLowerCase()).toList();
       // case insensitive comparison
       final uniqueNames = names.toSet();
-      if (uniqueNames.length < names.length) {
+      //t4y: in share by copy, always rename the items.to avoid the show dialog break the operation.
+      if (uniqueNames.length < names.length && moveType != MoveType.shareByCopy) {
         final value = await showDialog<NameConflictStrategy>(
           context: context,
           builder: (context) => AvesSingleSelectionDialog<NameConflictStrategy>(
             initialValue: nameConflictStrategy,
             options: Map.fromEntries(NameConflictStrategy.values.map((v) => MapEntry(v, v.getName(context)))),
-            message: originAlbums.length == 1 ? l10n.nameConflictDialogSingleSourceMessage : l10n.nameConflictDialogMultipleSourceMessage,
+            message: originAlbums.length == 1
+                ? l10n.nameConflictDialogSingleSourceMessage
+                : l10n.nameConflictDialogMultipleSourceMessage,
             confirmationButtonLabel: l10n.continueButtonLabel,
           ),
           routeSettings: const RouteSettings(name: AvesSingleSelectionDialog.routeName),
@@ -232,12 +278,22 @@ mixin EntryStorageMixin on FeedbackMixin, PermissionAwareMixin, SizeAwareMixin {
 
         // move
         final movedOps = successOps.where((v) => !v.skipped && !v.deleted).toSet();
-        final movedEntries = movedOps.map((v) => v.uri).map((uri) => entries.firstWhereOrNull((entry) => entry.uri == uri)).whereNotNull().toSet();
+        final movedEntries = movedOps
+            .map((v) => v.uri)
+            .map((uri) => entries.firstWhereOrNull((entry) => entry.uri == uri))
+            .whereNotNull()
+            .toSet();
+        //debugPrint('$runtimeType doQuickMove movedEntries:\n $movedEntries');
         await source.updateAfterMove(
           todoEntries: entries,
           moveType: moveType,
           destinationAlbums: destinationAlbums,
           movedOps: movedOps,
+          onUpdatedEntries: (entries) {
+            updatedEntries.addAll(entries);
+            debugPrint('$runtimeType doQuickMove entries:\n $entries');
+            debugPrint('$runtimeType doQuickMove updatedEntries:\n $updatedEntries');
+          },
         );
 
         // delete (when trying to move to bin obsolete entries)
@@ -247,8 +303,8 @@ mixin EntryStorageMixin on FeedbackMixin, PermissionAwareMixin, SizeAwareMixin {
 
         source.resumeMonitoring();
 
-        // cleanup
-        if ({MoveType.move, MoveType.toBin}.contains(moveType)) {
+        // cleanup// t4y: should not delete the share by copy dir.
+        if ({MoveType.move, MoveType.toBin}.contains(moveType) && !isShareByCopyDelete) {
           await storageService.deleteEmptyRegularDirectories(originAlbums);
         }
 
@@ -271,19 +327,22 @@ mixin EntryStorageMixin on FeedbackMixin, PermissionAwareMixin, SizeAwareMixin {
             final navigator = Navigator.maybeOf(context);
             if (toBin) {
               if (movedEntries.isNotEmpty) {
-                action = SnackBarAction(
-                  label: Themes.asButtonLabel(l10n.entryActionRestore),
-                  onPressed: () {
-                    if (navigator != null) {
-                      doMove(
-                        navigator.context,
-                        moveType: MoveType.fromBin,
-                        entries: movedEntries,
-                        hideShowAction: true,
-                      );
-                    }
-                  },
-                );
+                // t4y: share by copy not need to restore.
+                if (!isShareByCopyDelete) {
+                  action = SnackBarAction(
+                    label: Themes.asButtonLabel(l10n.entryActionRestore),
+                    onPressed: () {
+                      if (navigator != null) {
+                        doMove(
+                          navigator.context,
+                          moveType: MoveType.fromBin,
+                          entries: movedEntries,
+                          hideShowAction: true,
+                        );
+                      }
+                    },
+                  );
+                }
               }
             } else if (!hideShowAction) {
               action = SnackBarAction(
@@ -311,6 +370,23 @@ mixin EntryStorageMixin on FeedbackMixin, PermissionAwareMixin, SizeAwareMixin {
         }
       },
     );
+
+    if (moveType == MoveType.shareByCopy && settings.shareByCopyExpiredAutoRemove) {
+      // await deleteExpiredShareCopied(context);
+      final source = context.read<CollectionSource>();
+      final allCopiedEntries = source.allEntries
+          .where(
+              (entry) => {PathFilter(androidFileUtils.avesShareByCopyPath)}.any((f) => f.test(entry) && !entry.trashed))
+          .where(shareCopiedEntries.isShareCopied)
+          .toSet();
+      final todoEntries = switch (settings.shareByCopySetDateType) {
+        ShareByCopySetDateType.onlyThisTimeCopiedEntries =>
+          updatedEntries.isNotEmpty ? updatedEntries : allCopiedEntries,
+        ShareByCopySetDateType.allCopiedEntries => allCopiedEntries,
+      };
+      await EntrySetActionDelegate()
+          .setDateToNow(context, entries: todoEntries, showConfirm: false, isShareByCopy: true);
+    }
   }
 
   Future<void> doMove(
@@ -319,8 +395,12 @@ mixin EntryStorageMixin on FeedbackMixin, PermissionAwareMixin, SizeAwareMixin {
     required Set<AvesEntry> entries,
     bool hideShowAction = false,
     VoidCallback? onSuccess,
+    bool isShareByCopyDelete = false,
+    bool shareByCopyNeedRemove = true,
   }) async {
-    if (moveType == MoveType.toBin) {
+    final isViewerMode = context.read<ValueNotifier<AppMode>>().value == AppMode.view;
+    debugPrint('$runtimeType doMove in viewerMode?[$isViewerMode]:${context.read<ValueNotifier<AppMode>>().value}');
+    if (moveType == MoveType.toBin && !isShareByCopyDelete) {
       final l10n = context.l10n;
       if (!await showSkippableConfirmationDialog(
         context: context,
@@ -328,6 +408,12 @@ mixin EntryStorageMixin on FeedbackMixin, PermissionAwareMixin, SizeAwareMixin {
         message: l10n.binEntriesConfirmationDialogMessage(entries.length),
         confirmationButtonLabel: l10n.deleteButtonLabel,
       )) return;
+    }
+    // t4y delete the expired items first.
+    // when share by copy in viewer mode, not need to delete pre.
+    final needRemoveExpired = ((isViewerMode && settings.shareByCopyAppModeViewAutoRemove) || shareByCopyNeedRemove);
+    if (moveType == MoveType.shareByCopy && needRemoveExpired) {
+      await deleteExpiredShareCopied(context);
     }
 
     final entriesByDestination = <String, Set<AvesEntry>>{};
@@ -350,6 +436,19 @@ mixin EntryStorageMixin on FeedbackMixin, PermissionAwareMixin, SizeAwareMixin {
             entriesByDestination[originAlbum] = dirEntries.toSet();
           }
         });
+      case MoveType.shareByCopy:
+        final l10n = context.l10n;
+        final deleteTimeString = formatToLocalDuration(context, Duration(seconds: settings.shareByCopyRemoveInterval));
+        String dirName = pContext.basename(androidFileUtils.avesShareByCopyPath);
+        debugPrint('$runtimeType deleteTimeStrings:$deleteTimeString dirName:$dirName');
+
+        if (!await showSkippableConfirmationDialog(
+          context: context,
+          type: ConfirmationDialog.shareByCopy,
+          message: l10n.shareByCopyDialogMessage(androidFileUtils.avesShareByCopyPath, dirName, deleteTimeString),
+          confirmationButtonLabel: l10n.continueButtonLabel,
+        )) return;
+        entriesByDestination[androidFileUtils.avesShareByCopyPath] = entries;
     }
 
     await doQuickMove(
@@ -357,6 +456,7 @@ mixin EntryStorageMixin on FeedbackMixin, PermissionAwareMixin, SizeAwareMixin {
       moveType: moveType,
       entriesByDestination: entriesByDestination,
       onSuccess: onSuccess,
+      isShareByCopyDelete: isShareByCopyDelete,
     );
   }
